backend/rag_pipeline/RAG_BOT/requirements.txt
pinecone
openai
openpyxl
pypdf
python-dotenv
